LayoutTests/scrollingcoordinator/ios/scroll-element-at-point-expected.txt

PASS scrollTopAfterOutsideTopLeft is 0
PASS scrollTopAfterInsideTopLeft is 200
PASS scrollTopAfterOutsideBottomRight is 200
PASS scrollTopAfterInsideBottomRight is 400
PASS successfullyParsed is true

TEST COMPLETE
